refactor: rename sidebar
