+++ apps/web/src/app/routes/app/root.tsx
@@ -1,4 +1,4 @@
-import { Sidebar } from "@/components/layouts/side-bar";
+import { Sidebar } from "@/components/layouts/sidebar";
 
 export const AppRoot = () => {
     return (
